requirements.txt
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
gunicorn==20.1.0
jsonschema==4.6.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.4
pandas==1.4.3
pandas-datareader==0.10.0
pipenv==2022.10.25
plotly==5.11.0
python-dotenv==0.21.0
requests==2.28.1
requests-cache==0.9.7
requests-oauthlib==1.3.1
stack-data==0.3.0
virtualenv==20.16.6
virtualenv-clone==0.5.7